+++ src/components/tree/tree.common.ts
@@ -1,0 +1,14 @@
+import type IgcTreeItemComponent from './tree-item.js';
+
+export interface IgcTreeEventMap {
+  igcSelection: CustomEvent<{ newSelection: IgcTreeItemComponent[] }>;
+  igcItemExpanding: CustomEvent<IgcTreeItemComponent>;
+  igcItemExpanded: CustomEvent<IgcTreeItemComponent>;
+  igcItemCollapsing: CustomEvent<IgcTreeItemComponent>;
+  igcItemCollapsed: CustomEvent<IgcTreeItemComponent>;
+  igcActiveItem: CustomEvent<IgcTreeItemComponent>;
+}
+export interface IgcSelectionEventArgs {
+  detail: { readonly newSelection: IgcTreeItemComponent[] };
+  cancelable: boolean;
+}