+++ src/components/common/definitions/defineAllComponents.ts
@@ -41,6 +41,8 @@ import { defineComponents } from './defineComponents.js';
 import IgcCircularGradientComponent from '../../progress/circular-gradient.js';
 import IgcMaskInputComponent from '../../mask-input/mask-input.js';
 import IgcExpansionPanelComponent from '../../expansion-panel/expansion-panel.js';
+import IgcTreeComponent from '../../tree/tree.js';
+import IgcTreeItemComponent from '../../tree/tree-item.js';
 
 const allComponents: CustomElementConstructor[] = [
   IgcAvatarComponent,
@@ -77,6 +79,8 @@ const allComponents: CustomElementConstructor[] = [
   IgcRatingComponent,
   IgcRatingSymbolComponent,
   IgcRippleComponent,
+  IgcTreeComponent,
+  IgcTreeItemComponent,
   IgcSliderComponent,
   IgcToastComponent,
   IgcSliderLabelComponent,