+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Mask input [#173](https://github.com/IgniteUI/igniteui-webcomponents/issues/173)
 - Expansion Panel [#177](https://github.com/IgniteUI/igniteui-webcomponents/issues/177)
+- Tree [#188](https://github.com/IgniteUI/igniteui-webcomponents/issues/188)
+
 
 ## [3.1.0] - 2022-04-15
 ### Added